@@ -6,25 +6,22 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  automated_testing_framework_example: ^1.0.12
+  automated_testing_framework_example: ^2.0.0
   flutter:
     sdk: flutter
-  websafe_platform: ^1.0.5
 
-dependency_overrides:
-  automated_testing_framework:
-    path: ../
-  # automated_testing_framework_example:
-  #   path: ../../automated_testing_framework_example
-  # automated_testing_framework_models:
-  #   path: ../../automated_testing_framework_models
+# dependency_overrides:
+#   automated_testing_framework:
+#     path: ../
+#   automated_testing_framework_example:
+#     path: ../../automated_testing_framework_example
+#   automated_testing_framework_models:
+#     path: ../../automated_testing_framework_models
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.8.1
-  flutter_native_splash: ^0.1.9
   flutter_test:
     sdk: flutter
 
