@@ -1,30 +1,28 @@
 name: automated_testing_framework
 description: A testing framework that allows in app tests to be built and executed in an automated way.
-version: 2.0.0
+version: 3.0.0
 homepage: https://github.com/peiffer-innovations/automated_testing_framework
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  automated_testing_framework_models: ^1.1.0
-  child_builder: ^1.0.0+2
-  convert: ^2.1.1
-  device_info: ^1.0.0
+  automated_testing_framework_models: ^2.0.0+1
+  child_builder: ^2.0.0
+  convert: ^3.0.0
+  device_info: ^2.0.0
   flutter:
     sdk: flutter
-  form_validation: ^1.0.2+1
-  intl: ^0.16.1
-  json_class: ^1.0.7+2
-  json_theme: ^1.5.0+2
-  logging: ">=0.11.4 <2.0.0"
-  meta: ^1.2.4
-  package_info: ">=0.4.3+2 <2.0.0"
-  provider: ^4.3.2+4
-  static_translations: ^1.0.3+1
-  tinycolor: ^1.0.3
-  uuid: ^2.2.2
-  websafe_platform: ^1.1.0
+  form_validation: ^2.0.0
+  intl: ^0.17.0
+  json_class: ^2.0.0
+  json_theme: ^2.0.0+1
+  logging: ^1.0.0
+  meta: ^1.3.0
+  package_info: ^2.0.0
+  provider: ^5.0.0
+  static_translations: ^2.0.0
+  uuid: ^3.0.1
 
 dev_dependencies:
   flutter_test:
