@@ -18,10 +18,11 @@ import 'package:flutter/widgets.dart';
 /// are also cached.
 Iterable<Element> collectAllElementsFrom(
   Element rootElement, {
-  @required bool skipOffstage,
+  required bool skipOffstage,
 }) {
   return CachingIterable<Element>(
-      _DepthFirstChildIterator(rootElement, skipOffstage));
+    _DepthFirstChildIterator(rootElement, skipOffstage),
+  );
 }
 
 class _DepthFirstChildIterator implements Iterator<Element> {
@@ -30,7 +31,7 @@ class _DepthFirstChildIterator implements Iterator<Element> {
 
   final bool skipOffstage;
 
-  Element _current;
+  late Element _current;
 
   final List<Element> _stack;
 
@@ -49,8 +50,9 @@ class _DepthFirstChildIterator implements Iterator<Element> {
   }
 
   static Iterable<Element> _reverseChildrenOf(
-      Element element, bool skipOffstage) {
-    assert(element != null);
+    Element element,
+    bool skipOffstage,
+  ) {
     final children = <Element>[];
     if (skipOffstage) {
       element.debugVisitOnstageChildren(children.add);
