@@ -3,7 +3,8 @@
 // found in the LICENSE file.
 
 import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' show Tooltip;
+import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
 
 import 'all_elements.dart';
@@ -36,6 +37,21 @@ class CommonFinders {
   /// nodes that are [Offstage] or that are from inactive [Route]s.
   Finder text(String text, {bool skipOffstage = true}) =>
       _TextFinder(text, skipOffstage: skipOffstage);
+
+  /// Finds [Text] and [EditableText] widgets which contain the given
+  /// `pattern` argument.
+  ///
+  /// ## Sample code
+  ///
+  /// ```dart
+  /// expect(find.textContain('Back'), findsOneWidget);
+  /// expect(find.textContain(RegExp(r'(\w+)')), findsOneWidget);
+  /// ```
+  ///
+  /// If the `skipOffstage` argument is true (the default), then this skips
+  /// nodes that are [Offstage] or that are from inactive [Route]s.
+  Finder textContaining(Pattern pattern, {bool skipOffstage = true}) =>
+      _TextContainingFinder(pattern, skipOffstage: skipOffstage);
 
   /// Looks for widgets that contain a [Text] descendant with `text`
   /// in it.
@@ -191,7 +207,7 @@ class CommonFinders {
   /// If the `skipOffstage` argument is true (the default), then this skips
   /// nodes that are [Offstage] or that are from inactive [Route]s.
   Finder byWidgetPredicate(WidgetPredicate predicate,
-      {String description, bool skipOffstage = true}) {
+      {String? description, bool skipOffstage = true}) {
     return _WidgetPredicateFinder(predicate,
         description: description, skipOffstage: skipOffstage);
   }
@@ -235,7 +251,7 @@ class CommonFinders {
   /// If the `skipOffstage` argument is true (the default), then this skips
   /// nodes that are [Offstage] or that are from inactive [Route]s.
   Finder byElementPredicate(ElementPredicate predicate,
-      {String description, bool skipOffstage = true}) {
+      {String? description, bool skipOffstage = true}) {
     return _ElementPredicateFinder(predicate,
         description: description, skipOffstage: skipOffstage);
   }
@@ -257,8 +273,8 @@ class CommonFinders {
   /// If the [skipOffstage] argument is true (the default), then nodes that are
   /// [Offstage] or that are from inactive [Route]s are skipped.
   Finder descendant({
-    @required Finder of,
-    @required Finder matching,
+    required Finder of,
+    required Finder matching,
     bool matchRoot = false,
     bool skipOffstage = true,
   }) {
@@ -288,8 +304,8 @@ class CommonFinders {
   /// If the [matchRoot] argument is true then the widget(s) specified by [of]
   /// will be matched along with the ancestors.
   Finder ancestor({
-    @required Finder of,
-    @required Finder matching,
+    required Finder of,
+    required Finder matching,
     bool matchRoot = false,
   }) {
     return _AncestorFinder(of, matching, matchRoot: matchRoot);
@@ -311,13 +327,13 @@ class CommonFinders {
   /// ## Sample code
   ///
   /// ```dart
-  /// expect(find.BySemanticsLabel('Back'), findsOneWidget);
+  /// expect(find.bySemanticsLabel('Back'), findsOneWidget);
   /// ```
   ///
   /// If the `skipOffstage` argument is true (the default), then this skips
   /// nodes that are [Offstage] or that are from inactive [Route]s.
   Finder bySemanticsLabel(Pattern label, {bool skipOffstage = true}) {
-    if (WidgetsBinding.instance.pipelineOwner.semanticsOwner == null) {
+    if (WidgetsBinding.instance!.pipelineOwner.semanticsOwner == null) {
       throw StateError('Semantics are not enabled. '
           'Make sure to call tester.ensureSemantics() before using '
           'this finder, and call dispose on its return value after.');
@@ -329,7 +345,9 @@ class CommonFinders {
         if (element is! RenderObjectElement) {
           return false;
         }
-        final semanticsLabel = element.renderObject?.debugSemantics?.label;
+        // ignore: omit_local_variable_types
+        final String? semanticsLabel =
+            element.renderObject.debugSemantics?.label;
         if (semanticsLabel == null) {
           return false;
         }
@@ -375,12 +393,12 @@ abstract class Finder {
   @protected
   Iterable<Element> get allCandidates {
     return collectAllElementsFrom(
-      WidgetsBinding.instance.renderViewElement,
+      WidgetsBinding.instance!.renderViewElement!,
       skipOffstage: skipOffstage,
     );
   }
 
-  Iterable<Element> _cachedResult;
+  Iterable<Element>? _cachedResult;
 
   /// Returns the current result. If [precache] was called and returned true, this will
   /// cheaply return the result that was computed then. Otherwise, it creates a new
@@ -388,7 +406,8 @@ abstract class Finder {
   ///
   /// Calling this clears the cache from [precache].
   Iterable<Element> evaluate() {
-    final result = _cachedResult ?? apply(allCandidates);
+    // ignore: omit_local_variable_types
+    final Iterable<Element> result = _cachedResult ?? apply(allCandidates);
     _cachedResult = null;
     return result;
   }
@@ -431,12 +450,12 @@ abstract class Finder {
 
   @override
   String toString() {
-    final additional = skipOffstage ? ' (ignoring offstage widgets)' : '';
+    // ignore: omit_local_variable_types
+    final String additional =
+        skipOffstage ? ' (ignoring offstage widgets)' : '';
     final widgets = evaluate().toList();
     final count = widgets.length;
-    if (count == 0) {
-      return 'zero widgets with $description$additional';
-    }
+    if (count == 0) return 'zero widgets with $description$additional';
     if (count == 1) {
       return 'exactly one widget with $description$additional: ${widgets.single}';
     }
@@ -450,7 +469,7 @@ abstract class Finder {
 /// Applies additional filtering against a [parent] [Finder].
 abstract class ChainedFinder extends Finder {
   /// Create a Finder chained against the candidates of another [Finder].
-  ChainedFinder(this.parent) : assert(parent != null);
+  ChainedFinder(this.parent);
 
   /// Another [Finder] that will run first.
   final Finder parent;
@@ -521,11 +540,10 @@ class _HitTestableFinder extends ChainedFinder {
   @override
   Iterable<Element> filter(Iterable<Element> parentCandidates) sync* {
     for (final candidate in parentCandidates) {
-      final box = candidate.renderObject as RenderBox;
-      assert(box != null);
+      final box = candidate.renderObject! as RenderBox;
       final absoluteOffset = box.localToGlobal(alignment.alongSize(box.size));
       final hitResult = HitTestResult();
-      WidgetsBinding.instance.hitTest(hitResult, absoluteOffset);
+      WidgetsBinding.instance!.hitTest(hitResult, absoluteOffset);
       for (final entry in hitResult.path) {
         if (entry.target == candidate.renderObject) {
           yield candidate;
@@ -568,9 +586,33 @@ class _TextFinder extends MatchFinder {
     final widget = candidate.widget;
     if (widget is Text) {
       if (widget.data != null) return widget.data == text;
-      return widget.textSpan.toPlainText() == text;
+      assert(widget.textSpan != null);
+      return widget.textSpan!.toPlainText() == text;
     } else if (widget is EditableText) {
       return widget.controller.text == text;
+    }
+    return false;
+  }
+}
+
+class _TextContainingFinder extends MatchFinder {
+  _TextContainingFinder(this.pattern, {bool skipOffstage = true})
+      : super(skipOffstage: skipOffstage);
+
+  final Pattern pattern;
+
+  @override
+  String get description => 'text containing $pattern';
+
+  @override
+  bool matches(Element candidate) {
+    final widget = candidate.widget;
+    if (widget is Text) {
+      if (widget.data != null) return widget.data!.contains(pattern);
+      assert(widget.textSpan != null);
+      return widget.textSpan!.toPlainText().contains(pattern);
+    } else if (widget is EditableText) {
+      return widget.controller.text.contains(pattern);
     }
     return false;
   }
@@ -587,7 +629,16 @@ class _KeyFinder extends MatchFinder {
 
   @override
   bool matches(Element candidate) {
-    return candidate.widget.key == key;
+    var match = false;
+
+    var vk1 = candidate.widget.key is ValueKey
+        ? candidate.widget.key as ValueKey
+        : null;
+    var vk2 = key is ValueKey ? key as ValueKey : null;
+    if (candidate.widget.key is ValueKey) {
+      match = vk1 == vk2;
+    }
+    return match;
   }
 }
 
@@ -654,12 +705,12 @@ class _WidgetFinder extends MatchFinder {
 
 class _WidgetPredicateFinder extends MatchFinder {
   _WidgetPredicateFinder(this.predicate,
-      {String description, bool skipOffstage = true})
+      {String? description, bool skipOffstage = true})
       : _description = description,
         super(skipOffstage: skipOffstage);
 
   final WidgetPredicate predicate;
-  final String _description;
+  final String? _description;
 
   @override
   String get description =>
@@ -673,12 +724,12 @@ class _WidgetPredicateFinder extends MatchFinder {
 
 class _ElementPredicateFinder extends MatchFinder {
   _ElementPredicateFinder(this.predicate,
-      {String description, bool skipOffstage = true})
+      {String? description, bool skipOffstage = true})
       : _description = description,
         super(skipOffstage: skipOffstage);
 
   final ElementPredicate predicate;
-  final String _description;
+  final String? _description;
 
   @override
   String get description =>
@@ -704,9 +755,9 @@ class _DescendantFinder extends Finder {
 
   @override
   String get description {
-    if (matchRoot) {
+    if (matchRoot)
+      // ignore: curly_braces_in_flow_control_structures
       return '${descendant.description} in the subtree(s) beginning with ${ancestor.description}';
-    }
     return '${descendant.description} that has ancestor(s) with ${ancestor.description}';
   }
 
@@ -718,8 +769,10 @@ class _DescendantFinder extends Finder {
 
   @override
   Iterable<Element> get allCandidates {
-    final ancestorElements = ancestor.evaluate();
-    final candidates = ancestorElements
+    // ignore: omit_local_variable_types
+    final Iterable<Element> ancestorElements = ancestor.evaluate();
+    // ignore: omit_local_variable_types
+    final List<Element> candidates = ancestorElements
         .expand<Element>((Element element) =>
             collectAllElementsFrom(element, skipOffstage: skipOffstage))
         .toSet()
@@ -739,9 +792,9 @@ class _AncestorFinder extends Finder {
 
   @override
   String get description {
-    if (matchRoot) {
+    if (matchRoot)
+      // ignore: curly_braces_in_flow_control_structures
       return 'ancestor ${ancestor.description} beginning with ${descendant.description}';
-    }
     return '${ancestor.description} which is an ancestor of ${descendant.description}';
   }
 
@@ -753,9 +806,11 @@ class _AncestorFinder extends Finder {
 
   @override
   Iterable<Element> get allCandidates {
-    final candidates = <Element>[];
+    // ignore: omit_local_variable_types
+    final List<Element> candidates = <Element>[];
     for (final root in descendant.evaluate()) {
-      final ancestors = <Element>[];
+      // ignore: omit_local_variable_types
+      final List<Element> ancestors = <Element>[];
       if (matchRoot) ancestors.add(root);
       root.visitAncestorElements((Element element) {
         ancestors.add(element);
