@@ -23,10 +23,7 @@ class TestPointer {
     this.kind = PointerDeviceKind.touch,
     this._device,
     int buttons = kPrimaryButton,
-  ])  : assert(kind != null),
-        assert(pointer != null),
-        assert(buttons != null),
-        _buttons = buttons {
+  ]) : _buttons = buttons {
     switch (kind) {
       case PointerDeviceKind.mouse:
         _device ??= 1;
@@ -44,8 +41,8 @@ class TestPointer {
   ///
   /// Set when the object is constructed. Defaults to 1 if the [kind] is
   /// [PointerDeviceKind.mouse], and 0 otherwise.
-  int get device => _device;
-  int _device;
+  int? get device => _device;
+  int? _device;
 
   /// The pointer identifier used for events generated by this object.
   ///
@@ -72,15 +69,15 @@ class TestPointer {
   /// The position of the last event sent by this object.
   ///
   /// If no event has ever been sent by this object, returns null.
-  Offset get location => _location;
-  Offset _location;
+  Offset? get location => _location;
+  Offset? _location;
 
   /// If a custom event is created outside of this class, this function is used
   /// to set the [isDown].
   bool setDownInfo(
     PointerEvent event,
     Offset newLocation, {
-    int buttons,
+    int? buttons,
   }) {
     _location = newLocation;
     if (buttons != null) _buttons = buttons;
@@ -110,7 +107,7 @@ class TestPointer {
   PointerDownEvent down(
     Offset newLocation, {
     Duration timeStamp = Duration.zero,
-    int buttons,
+    int? buttons,
   }) {
     assert(!isDown);
     _isDown = true;
@@ -119,9 +116,9 @@ class TestPointer {
     return PointerDownEvent(
       timeStamp: timeStamp,
       kind: kind,
-      device: _device,
+      device: _device!,
       pointer: pointer,
-      position: location,
+      position: location!,
       buttons: _buttons,
     );
   }
@@ -139,20 +136,20 @@ class TestPointer {
   PointerMoveEvent move(
     Offset newLocation, {
     Duration timeStamp = Duration.zero,
-    int buttons,
+    int? buttons,
   }) {
     assert(
         isDown,
         'Move events can only be generated when the pointer is down. To '
         'create a movement event simulating a pointer move when the pointer is '
         'up, use hover() instead.');
-    final delta = newLocation - location;
+    final delta = newLocation - location!;
     _location = newLocation;
     if (buttons != null) _buttons = buttons;
     return PointerMoveEvent(
       timeStamp: timeStamp,
       kind: kind,
-      device: _device,
+      device: _device!,
       pointer: pointer,
       position: newLocation,
       delta: delta,
@@ -172,9 +169,9 @@ class TestPointer {
     return PointerUpEvent(
       timeStamp: timeStamp,
       kind: kind,
-      device: _device,
+      device: _device!,
       pointer: pointer,
-      position: location,
+      position: location!,
     );
   }
 
@@ -190,9 +187,9 @@ class TestPointer {
     return PointerCancelEvent(
       timeStamp: timeStamp,
       kind: kind,
-      device: _device,
+      device: _device!,
       pointer: pointer,
-      position: location,
+      position: location!,
     );
   }
 
@@ -203,14 +200,13 @@ class TestPointer {
   /// specific time stamp by passing the `timeStamp` argument.
   PointerAddedEvent addPointer({
     Duration timeStamp = Duration.zero,
-    Offset location,
+    Offset? location,
   }) {
-    assert(timeStamp != null);
     _location = location ?? _location;
     return PointerAddedEvent(
       timeStamp: timeStamp,
       kind: kind,
-      device: _device,
+      device: _device!,
       position: _location ?? Offset.zero,
     );
   }
@@ -222,14 +218,13 @@ class TestPointer {
   /// specific time stamp by passing the `timeStamp` argument.
   PointerRemovedEvent removePointer({
     Duration timeStamp = Duration.zero,
-    Offset location,
+    Offset? location,
   }) {
-    assert(timeStamp != null);
     _location = location ?? _location;
     return PointerRemovedEvent(
       timeStamp: timeStamp,
       kind: kind,
-      device: _device,
+      device: _device!,
       position: _location ?? Offset.zero,
     );
   }
@@ -245,20 +240,18 @@ class TestPointer {
     Offset newLocation, {
     Duration timeStamp = Duration.zero,
   }) {
-    assert(newLocation != null);
-    assert(timeStamp != null);
     assert(
         !isDown,
         'Hover events can only be generated when the pointer is up. To '
         'simulate movement when the pointer is down, use move() instead.');
     assert(kind != PointerDeviceKind.touch,
         "Touch pointers can't generate hover events");
-    final delta = location != null ? newLocation - location : Offset.zero;
+    final delta = location != null ? newLocation - location! : Offset.zero;
     _location = newLocation;
     return PointerHoverEvent(
       timeStamp: timeStamp,
       kind: kind,
-      device: _device,
+      device: _device!,
       position: newLocation,
       delta: delta,
     );
@@ -273,15 +266,13 @@ class TestPointer {
     Offset scrollDelta, {
     Duration timeStamp = Duration.zero,
   }) {
-    assert(scrollDelta != null);
-    assert(timeStamp != null);
     assert(kind != PointerDeviceKind.touch,
         "Touch pointers can't generate pointer signal events");
     return PointerScrollEvent(
       timeStamp: timeStamp,
       kind: kind,
-      device: _device,
-      position: location,
+      device: _device!,
+      position: location!,
       scrollDelta: scrollDelta,
     );
   }
@@ -290,7 +281,7 @@ class TestPointer {
 /// Signature for a callback that can dispatch events and returns a future that
 /// completes when the event dispatch is complete.
 typedef EventDispatcher = Future<void> Function(
-    PointerEvent event, HitTestResult result);
+    PointerEvent event, HitTestResult? result);
 
 /// Signature for callbacks that perform hit-testing at a given location.
 typedef HitTester = HitTestResult Function(Offset location);
@@ -318,18 +309,13 @@ class TestGesture {
   /// None of the arguments may be null. The `dispatcher` and `hitTester`
   /// arguments are required.
   TestGesture({
-    @required EventDispatcher dispatcher,
-    @required HitTester hitTester,
+    required EventDispatcher dispatcher,
+    required HitTester hitTester,
     int pointer = 1,
     PointerDeviceKind kind = PointerDeviceKind.touch,
-    int device,
+    int? device,
     int buttons = kPrimaryButton,
-  })  : assert(dispatcher != null),
-        assert(hitTester != null),
-        assert(pointer != null),
-        assert(kind != null),
-        assert(buttons != null),
-        _dispatcher = dispatcher,
+  })  : _dispatcher = dispatcher,
         _hitTester = hitTester,
         _pointer = TestPointer(pointer, kind, device, buttons),
         _result = null;
@@ -355,7 +341,7 @@ class TestGesture {
   final EventDispatcher _dispatcher;
   final HitTester _hitTester;
   final TestPointer _pointer;
-  HitTestResult _result;
+  HitTestResult? _result;
 
   /// In a test, send a move event that moves the pointer by the given offset.
   @visibleForTesting
@@ -367,7 +353,7 @@ class TestGesture {
 
   /// In a test, send a pointer add event for this pointer.
   Future<void> addPointer(
-      {Duration timeStamp = Duration.zero, Offset location}) {
+      {Duration timeStamp = Duration.zero, Offset? location}) {
     return _dispatcher(
         _pointer.addPointer(
             timeStamp: timeStamp, location: location ?? _pointer.location),
@@ -376,7 +362,7 @@ class TestGesture {
 
   /// In a test, send a pointer remove event for this pointer.
   Future<void> removePointer(
-      {Duration timeStamp = Duration.zero, Offset location}) {
+      {Duration timeStamp = Duration.zero, Offset? location}) {
     return _dispatcher(
         _pointer.removePointer(
             timeStamp: timeStamp, location: location ?? _pointer.location),
@@ -389,7 +375,7 @@ class TestGesture {
   /// up, then a hover event is dispatched. Touch devices are not able to send
   /// hover events.
   Future<void> moveBy(Offset offset, {Duration timeStamp = Duration.zero}) {
-    return moveTo(_pointer.location + offset, timeStamp: timeStamp);
+    return moveTo(_pointer.location! + offset, timeStamp: timeStamp);
   }
 
   /// Send a move event moving the pointer to the given location.
