@@ -7,12 +7,9 @@ class ProgressValue {
   /// Creates the progress instance.
   ProgressValue({
     this.error = false,
-    @required this.max,
-    @required int value,
-  })  : assert(error != null),
-        assert(max != null),
-        assert(value != null),
-        assert(max > 0),
+    required this.max,
+    required int value,
+  })   : assert(max > 0),
         assert(value >= 0),
         value = min(value, max);
 
