@@ -1,20 +1,19 @@
 import 'package:automated_testing_framework/automated_testing_framework.dart';
 import 'package:json_class/json_class.dart';
-import 'package:meta/meta.dart';
 
 /// Step that double-taps a [Testable] widget.
 class DoubleTapStep extends TestRunnerStep {
   DoubleTapStep({
-    @required this.testableId,
+    required this.testableId,
     this.timeout,
   }) : assert(testableId?.isNotEmpty == true);
 
   /// The id of the [Testable] widget to interact with.
-  final String testableId;
+  final String? testableId;
 
   /// The maximum amount of time this step will wait while searching for the
   /// [Testable] on the widget tree.
-  final Duration timeout;
+  final Duration? timeout;
 
   /// Creates an instance from a JSON-like map structure.  This expects the
   /// following format:
@@ -28,8 +27,8 @@ class DoubleTapStep extends TestRunnerStep {
   ///
   /// See also:
   /// * [JsonClass.parseDurationFromSeconds]
-  static DoubleTapStep fromDynamic(dynamic map) {
-    DoubleTapStep result;
+  static DoubleTapStep? fromDynamic(dynamic map) {
+    DoubleTapStep? result;
 
     if (map != null) {
       result = DoubleTapStep(
@@ -45,11 +44,11 @@ class DoubleTapStep extends TestRunnerStep {
   /// then will attempt to double tap the widget on center point of the widget.
   @override
   Future<void> execute({
-    @required CancelToken cancelToken,
-    @required TestReport report,
-    @required TestController tester,
+    required CancelToken cancelToken,
+    required TestReport report,
+    required TestController tester,
   }) async {
-    String testableId = tester.resolveVariable(this.testableId);
+    String? testableId = tester.resolveVariable(this.testableId);
     assert(testableId?.isNotEmpty == true);
 
     var name = "double_tap('$testableId')";
