@@ -1,6 +1,5 @@
 import 'package:automated_testing_framework/automated_testing_framework.dart';
 import 'package:json_class/json_class.dart';
-import 'package:meta/meta.dart';
 
 /// Step that sends a longPress event to a [Testable] widget.
 ///
@@ -8,16 +7,16 @@ import 'package:meta/meta.dart';
 /// * [WidgetController.longPress]
 class LongPressStep extends TestRunnerStep {
   LongPressStep({
-    @required this.testableId,
+    required this.testableId,
     this.timeout,
   }) : assert(testableId?.isNotEmpty == true);
 
   /// The id of the [Testable] widget to interact with.
-  final String testableId;
+  final String? testableId;
 
   /// The maximum amount of time this step will wait while searching for the
   /// [Testable] on the widget tree.
-  final Duration timeout;
+  final Duration? timeout;
 
   /// Creates an instance from a JSON-like map structure.  This expects the
   /// following format:
@@ -31,8 +30,8 @@ class LongPressStep extends TestRunnerStep {
   ///
   /// See also:
   /// * [JsonClass.parseDurationFromSeconds]
-  static LongPressStep fromDynamic(dynamic map) {
-    LongPressStep result;
+  static LongPressStep? fromDynamic(dynamic map) {
+    LongPressStep? result;
 
     if (map != null) {
       result = LongPressStep(
@@ -48,11 +47,11 @@ class LongPressStep extends TestRunnerStep {
   /// then will attempt to long press the widget on center point of the widget.
   @override
   Future<void> execute({
-    @required CancelToken cancelToken,
-    @required TestReport report,
-    @required TestController tester,
+    required CancelToken cancelToken,
+    required TestReport report,
+    required TestController tester,
   }) async {
-    String testableId = tester.resolveVariable(this.testableId);
+    String? testableId = tester.resolveVariable(this.testableId);
     assert(testableId?.isNotEmpty == true);
 
     var name = "long_press('$testableId')";
