@@ -1,15 +1,12 @@
 import 'package:automated_testing_framework/automated_testing_framework.dart';
-import 'package:flutter/material.dart';
-import 'package:meta/meta.dart';
 
 /// Sets a variable on the identified [TestController].
 class SetVariableStep extends TestRunnerStep {
   SetVariableStep({
     String type = 'String',
-    @required this.value,
-    @required this.variableName,
-  })  : assert(type != null),
-        assert(type == 'bool' ||
+    required this.value,
+    required this.variableName,
+  })   : assert(type == 'bool' ||
             type == 'double' ||
             type == 'int' ||
             type == 'String'),
@@ -24,10 +21,10 @@ class SetVariableStep extends TestRunnerStep {
   final String type;
 
   /// The string representation of the value to set.
-  final String value;
+  final String? value;
 
   /// The variable name of the variable to set on the controller.
-  final String variableName;
+  final String? variableName;
 
   /// Creates an instance from a JSON-like map structure.  This expects the
   /// following format:
@@ -39,8 +36,8 @@ class SetVariableStep extends TestRunnerStep {
   ///   "variableName": <String>,
   /// }
   /// ```
-  static SetVariableStep fromDynamic(dynamic map) {
-    SetVariableStep result;
+  static SetVariableStep? fromDynamic(dynamic map) {
+    SetVariableStep? result;
 
     if (map != null) {
       result = SetVariableStep(
@@ -58,20 +55,19 @@ class SetVariableStep extends TestRunnerStep {
   /// Sets the variable on the [TestController].
   @override
   Future<void> execute({
-    @required CancelToken cancelToken,
-    @required TestReport report,
-    @required TestController tester,
+    required CancelToken cancelToken,
+    required TestReport report,
+    required TestController tester,
   }) async {
     String type = tester.resolveVariable(this.type);
-    String value = tester.resolveVariable(this.value);
+    String? value = tester.resolveVariable(this.value);
     String variableName = tester.resolveVariable(this.variableName);
 
-    assert(type != null);
     assert(type == 'bool' ||
         type == 'double' ||
         type == 'int' ||
         type == 'String');
-    assert(variableName?.isNotEmpty == true);
+    assert(variableName.isNotEmpty == true);
     var name = "set_variable('$variableName', '$type', '$value')";
 
     log(
