@@ -1,13 +1,11 @@
 import 'package:automated_testing_framework/automated_testing_framework.dart';
 import 'package:json_class/json_class.dart';
-import 'package:meta/meta.dart';
 
 /// Simple test step that will sleep for a given period of time.
 class SleepStep extends TestRunnerStep {
   SleepStep({
-    @required this.timeout,
-  })  : assert(timeout != null),
-        assert(timeout.inMilliseconds >= 0);
+    required this.timeout,
+  }) : assert(timeout.inMilliseconds >= 0);
 
   /// The maximum amount of time to sleep for.
   final Duration timeout;
@@ -23,12 +21,12 @@ class SleepStep extends TestRunnerStep {
   ///
   /// See also:
   /// * [JsonClass.parseDurationFromSeconds]
-  static SleepStep fromDynamic(dynamic map) {
-    SleepStep result;
+  static SleepStep? fromDynamic(dynamic map) {
+    SleepStep? result;
 
     if (map != null) {
       result = SleepStep(
-        timeout: JsonClass.parseDurationFromSeconds(map['timeout']),
+        timeout: JsonClass.parseDurationFromSeconds(map['timeout'])!,
       );
     }
 
@@ -38,9 +36,9 @@ class SleepStep extends TestRunnerStep {
   /// Simply sleeps for the time specified by [timeout].
   @override
   Future<void> execute({
-    CancelToken cancelToken,
-    TestReport report,
-    @required TestController tester,
+    CancelToken? cancelToken,
+    TestReport? report,
+    required TestController tester,
   }) async {
     var name = "sleep('${timeout.inMilliseconds}')";
     log(
