@@ -8,10 +8,10 @@ import 'package:flutter/material.dart';
 /// be given the a new [ValueKey] with the value from the child's key prefixed
 /// with `value_key_`.
 Widget testableChildBuilder(BuildContext context, Widget child) {
-  String id;
+  String? id;
 
   if (child.key != null && child.key is ValueKey) {
-    ValueKey childKey = child.key;
+    var childKey = child.key as ValueKey<dynamic>;
 
     if (childKey.value is String) {
       id = 'value_key_${childKey.value}';
