@@ -1,9 +1,10 @@
+import 'dart:io';
+
 import 'package:automated_testing_framework/automated_testing_framework.dart';
 import 'package:automated_testing_framework_example/automated_testing_framework_example.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
-import 'package:websafe_platform/websafe_platform.dart';
 
 void main() {
   TestAppSettings.initialize(appIdentifier: 'ATF Core');
@@ -22,12 +23,11 @@ void main() {
   });
 
   var gestures = TestableGestures();
-  var wsPlatform = WebsafePlatform();
-  if (wsPlatform.isFuchsia() ||
-      wsPlatform.isLinux() ||
-      wsPlatform.isMacOS() ||
-      wsPlatform.isWindows() ||
-      wsPlatform.isWeb()) {
+  if (kIsWeb ||
+      Platform.isFuchsia ||
+      Platform.isLinux ||
+      Platform.isMacOS ||
+      Platform.isWindows) {
     gestures = TestableGestures(
       widgetLongPress: null,
       widgetSecondaryLongPress: TestableGestureAction.open_test_actions_page,
