@@ -15,7 +15,7 @@ class DoubleTapForm extends TestStepForm {
   @override
   Widget buildForm(
     BuildContext context,
-    Map<String, dynamic> values, {
+    Map<String, dynamic>? values, {
     bool minify = false,
   }) {
     return Column(
@@ -37,7 +37,7 @@ class DoubleTapForm extends TestStepForm {
               validators: [
                 RequiredValidator(),
               ],
-              values: values,
+              values: values!,
             ),
             if (minify != true) ...[
               SizedBox(height: 16.0),
