@@ -16,7 +16,7 @@ class ScrollUntilVisibleForm extends TestStepForm {
   @override
   Widget buildForm(
     BuildContext context,
-    Map<String, dynamic> values, {
+    Map<String, dynamic>? values, {
     bool minify = false,
   }) {
     return Column(
@@ -38,7 +38,7 @@ class ScrollUntilVisibleForm extends TestStepForm {
               validators: [
                 RequiredValidator(),
               ],
-              values: values,
+              values: values!,
             ),
             SizedBox(height: 16.0),
             buildEditText(
